refactored code

+++ api/admin/views.py
@@ -3,28 +3,25 @@ import random
 import string
 from ..utils import db
 from functools import wraps
-from flask import request
+from flask import request, abort
 from flask_restx import Resource, fields, Namespace
 from ..auth.views import generate_random_string, generate_password
-from ..models.user import Student, User, Tutor
-from ..models.courses import Course, StudentCourse, Score
+from ..models.user import Student, User
+from ..models.courses import Course, StudentCourse
 from http import HTTPStatus
 from flask_jwt_extended import jwt_required, get_jwt_identity, unset_jwt_cookies
 
 
 
-admin_namespace = Namespace('student', description = 'Admin accessible route')
+admin_namespace = Namespace('Admin', description = 'Admin accessible route')
 
-def is_admin(self, f):
+def is_admin(f):
     @wraps(f)
     def decorated_function(*args, **kwargs):
         # check if user is admin
-        user_id = get_jwt_identity()
-        current_user = User.get_by_id('user_id')
-        if not current_user.is_admin:
-            return {
-                'message': 'You are not authorized to perform this action'
-            }
+        password = get_jwt_identity()
+        if password == 'Student101':
+            abort(401, 'You are not authorized to access this route')
         return f(*args, **kwargs)
     return decorated_function
 
@@ -42,29 +39,46 @@ def calculate_grades(self):
         return 'F'
 
 
-course_model = admin_namespace.model(
+create_course_model = admin_namespace.model(
     'Course', {
         'id': fields.String(required=True),
-        'course_title': fields.String(required=True, description="Course Name"),
-        'course_code': fields.String(required=True, description="'Course Code"),
-        'course_description': fields.String(required=True, description="Course Description"),
-        'course_unit': fields.String(required=True, description="Course Unit"),
-        'course_level': fields.String(required=True, description="Course Level"),
-        'score': fields.Float(required=True, default=0.0),
+        'course_code': fields.String(description="'Course Code"),
+        'course_unit': fields.String(description="Course Unit"),
+        'course_title': fields.String( description="Course Name"),
         'tutor_name': fields.String(required=True, description="Course Level"),
-        
+        'score': fields.Float(required=True, default=0.0),
+        'grade': fields.String(required=True, default='N/A')
     }
 )
+
 
 
 
 course_list_model = admin_namespace.model(
     'CourseList', {
-        'courses': fields.List(fields.Nested(course_model))
+        'courses': fields.List(fields.Nested(create_course_model))
     }
 )
 
-student_score = admin_namespace.model('StudentScores', {
+course_model = admin_namespace.model(
+    'Course', {
+        'course_id': fields.String(required=True),
+        }
+)
+
+
+show_course_model = admin_namespace.model(
+    'ShowCourse', {
+        'id': fields.String(required=True),
+        'course_title': fields.String(required=True, description="Course Name"),
+        'course_code': fields.String(required=True, description="'Course Code"),
+        'course_unit': fields.String(required=True, description="Course Unit"),
+        'score': fields.Float(required=True, default=0.0),
+        'tutor_name': fields.String(required=True, description="Course Level")
+    }
+)
+
+student_score = admin_namespace.model('StudentScore', {
     'score': fields.Integer
 })
 
@@ -84,29 +98,36 @@ student_model = admin_namespace.model(
 	}
 )
 
-
-
-
-course_model = admin_namespace.model(
-    'Course', {
-		'id': fields.String(required=True),
-        'course_title': fields.String(required=True, description="Course Name"),
-        'course_code': fields.String(required=True, description="'Course Code"),
-        'course_unit': fields.String(required=True, description="Course Unit"),
-        'tutor_name': fields.String(required=True, description="Course Level")
-	}
+simple_student_model = admin_namespace.model(
+    'SimpleStudent', {
+        'id': fields.String(dump_only=True),
+        'first_name': fields.String(required=True, description="Student's First Name"),
+        'last_name': fields.String(required=True, description="'Student's Last Name"),
+        'email': fields.String(required=True, description="Student's Email"),
+        'password': fields.String(required=True, default="Student101", description="Student's Password"),
+    }
 )
-
 
 
 
 student_course_model = admin_namespace.model(
     'StudentCourse', {
-        'id': fields.String(required=True, description="'User's Name"),
-        'course_id': fields.String(required=True, description="Student's First Name"),
-        'student_id': fields.String(required=True, description="'Student's Last Name")
-    }
+		'id': fields.String(required=True),
+        'course_code': fields.String(description="'Course Code"),
+        'course_unit': fields.String( description="Course Name"),
+        'score': fields.Float(required=True, default=0.0),
+        'grade': fields.String(required=True, default='N/A'),
+        'first_name': fields.String(required=True, description="Student's First Name"),
+        'last_name': fields.String(required=True, description="'Student's Last Name"),
+        'student_id': fields.String(required=True, description="Studend ID"),
+        'course_id': fields.String(required=True, description="Course ID"),
+
+
+	}
 )
+
+
+
 
 score_model = admin_namespace.model(
     'Score', {
@@ -130,6 +151,7 @@ class GetStudents(Resource):
         description='Get all students',
     )
     @jwt_required()
+    @is_admin
     def get(self):
         """
             Get all students
@@ -145,34 +167,31 @@ class GetStudents(Resource):
         return students, HTTPStatus.OK
     
 
-    @admin_namespace.expect(student_model)
-    @admin_namespace.marshal_with(student_model)
+    @admin_namespace.expect(simple_student_model)
+    @admin_namespace.marshal_with(simple_student_model)
     @admin_namespace.doc(description="Create a new student")
     @jwt_required()
+    @is_admin
     def post(self):
         """
         Create a new student
         """
         data = admin_namespace.payload
 
-        student = Student.query.filter_by(student_id=data['student_id']).first()
+        # student = Student.query.filter_by(student_id=data['student_id']).first()
+        existing_email = Student.query.filter_by(email=data['email']).first()
 
-        # courses = Course.query.filter(Course.id.in_(data['courses'])).all()
-
-
-        if student:
+        if existing_email:
             return {
-                'message': 'This student already exists'
+                'message': 'This email already exists'
             }, HTTPStatus.BAD_REQUEST
+
 
         new_student = Student(
             student_id='ALT00' + generate_random_string(1),
             email=data['email'],
             first_name=data['first_name'],
-            last_name=data['last_name'],
-            password_hash=data['password'],
-            # registered_courses=courses, # set to the list of courses
-            gpa=data['gpa']
+            last_name=data['last_name']
         )
         new_student.save()
 
@@ -194,6 +213,7 @@ class GetUpdateDeleteStudent(Resource):
         }
     )
     @jwt_required()
+    @is_admin
     def get(self, student_id):
         """
             Retrieve a student by id
@@ -209,33 +229,7 @@ class GetUpdateDeleteStudent(Resource):
         return student, HTTPStatus.OK
     
 
-    @admin_namespace.expect(student_model)
-    @admin_namespace.marshal_with(student_model)
-    @admin_namespace.doc(
-        description='Update a student by id', params={
-            'student_id': 'The student id'
-        }
-    )
-    @jwt_required()
-    def put(self, student_id):
-        """
-            Update a student by id
-        """
 
-        update_student = Student.get_by_id(student_id)
-        data = admin_namespace.payload
-
-        update_student.first_name = data['first_name']
-        update_student.last_name = data['last_name']
-        update_student.password = data['password']
-        update_student.registered_courses = data['registered_courses']
-        update_student.gpa = data['gpa']
-
-        update_student.update()
-
-        return update_student, HTTPStatus.OK, {
-            'message': 'Student updated successfully'
-        }
     
 
     @admin_namespace.expect(student_model)
@@ -246,6 +240,7 @@ class GetUpdateDeleteStudent(Resource):
         }  
     )
     @jwt_required()
+    @is_admin
     def delete(self, student_id):
         """
             Delete a student by id
@@ -265,11 +260,12 @@ class GetUpdateDeleteStudent(Resource):
 @admin_namespace.route('/courses')
 class GetCourses(Resource):
     
-    @admin_namespace.marshal_with(course_model, as_list=True)
+    @admin_namespace.marshal_with(create_course_model, as_list=True)
     @admin_namespace.doc(
         description='Get all registered courses',
     )
     @jwt_required()
+    @is_admin
     def get(self):
         """
             Retrieve all courses
@@ -285,13 +281,13 @@ class GetCourses(Resource):
         return courses, HTTPStatus.OK
     
 
-    @admin_namespace.expect(course_model)
-    @admin_namespace.marshal_with(course_model)
+    @admin_namespace.expect(create_course_model)
+    @admin_namespace.marshal_with(create_course_model)
     @admin_namespace.doc(
         description="Create a new course"
     )
     @jwt_required()
-    # @is_admin()
+    @is_admin
     def post(Self):
         """
             Create a new course
@@ -323,33 +319,6 @@ class GetCourses(Resource):
     
 
 
-@admin_namespace.route('/students/course/<int:course_id>')
-class GetUpdateDelete(Resource):
-    @admin_namespace.marshal_with(student_model)
-    @admin_namespace.doc(
-        description='Retrieve all students registered for a particular course', params={
-            'course_id': 'The course id'
-        }
-    )
-    @jwt_required()
-    def get(self, course_id):
-        """
-            Retrieve all students registered for a particular course
-        """
-        courses = Course.query.filter_by(id=course_id).all()
-        if courses:
-            students = Student.query.all()
-            if students:
-                return students, HTTPStatus.OK
-            return {
-                'message': 'No student has been registered for this course'
-            }
-        
-        return {
-            'message': 'Course does not exist'
-        }, HTTPStatus.BAD_REQUEST
-    
-
 
 @admin_namespace.route('/course/<int:course_id>/students')
 class GetStudent(Resource):
@@ -360,6 +329,7 @@ class GetStudent(Resource):
         }
     )
     @jwt_required()
+    @is_admin
     def get(self, course_id):
         """
             Retrieve all students for a particular course
@@ -367,29 +337,26 @@ class GetStudent(Resource):
         courses = StudentCourse.query.filter_by(id=course_id).all()
         for course in courses:
             students = Student.query.all()
-            if students in courses:
+            if students not in courses:
                 return students, HTTPStatus.OK
             return {
                 'message': 'No student has been registered for this course'
             }
+        return {
+            'message': 'Course does not exist'
+        }
 
 
 
-
-
-            # students = []
-            # for course in courses:
-            #     student = Student.query.filter_by(id=course.student_id).first()
-            #     students.append(student)
 
 
     
 
 
-@admin_namespace.route('/course/<int:course_id>')
-class GetUpdateDeleteCourse(Resource):
+@admin_namespace.route('/course/<int:course_id>/students/<int:student_id>')
+class GetCourse(Resource):
     
-    @admin_namespace.marshal_with(course_model)
+    @admin_namespace.marshal_with(create_course_model)
     @admin_namespace.doc(
         description='Get a course by id', params={
             'course_id': 'The course id',
@@ -397,6 +364,7 @@ class GetUpdateDeleteCourse(Resource):
         }
     )
     @jwt_required()
+    @is_admin
     def get(self, course_id):
         """
             Retrieve a course by id
@@ -412,14 +380,15 @@ class GetUpdateDeleteCourse(Resource):
         return course, HTTPStatus.OK
     
 
-    @admin_namespace.expect(course_model)
-    @admin_namespace.marshal_with(course_model)
+    @admin_namespace.expect(create_course_model)
+    @admin_namespace.marshal_with(create_course_model)
     @admin_namespace.doc(
         description='Update a course by id', params={
             'course_id': 'The course id'
         }
     )
     @jwt_required()
+    @is_admin
     def patch(self, course_id, student_id):
         """
             Update a student's course by id
@@ -429,9 +398,12 @@ class GetUpdateDeleteCourse(Resource):
         
 
         if student:
+            # get the course
+            course = Course.query.filter_by(id=course_id).first()
+                
             # student_id = Student.get_by_id(student_id)
             # check if student is already registered for the course
-            if course_id in student.registered_courses:
+            if course in student.registered_courses:
                 update_course = Course.get_by_id(course_id)
                 data = admin_namespace.payload
 
@@ -451,20 +423,24 @@ class GetUpdateDeleteCourse(Resource):
                 }
     
 
-    @admin_namespace.expect(course_model)
-    @admin_namespace.marshal_with(course_model)
+        
     @admin_namespace.doc(
         description='Delete a course by id', params={
             'course_id': 'The course id'
         }  
     )
     @jwt_required()
+    @is_admin
     def delete(self, course_id):
         """
             Delete a course by id
         """
 
         course = Course.get_by_id(course_id)
+        if course is None:
+            return {
+                'message': 'This course does not exist'
+            }, HTTPStatus.BAD_REQUEST
 
         course.delete()
 
@@ -480,43 +456,45 @@ class GetUpdateDeleteCourse(Resource):
 
 
 
-@admin_namespace.route('/course/<int:course_id>/student/<int:student_id>')
+@admin_namespace.route('/student/<int:student_id>')
 class RegisterStudentCourse(Resource):
     
-    @admin_namespace.expect(course_model)
-    # @admin_namespace.marshal_with(student_course_model)
+    @admin_namespace.expect(student_course_model)
+    @admin_namespace.marshal_with(student_course_model)
     @admin_namespace.doc(
         description='Register a student for a course'
     )
     @jwt_required()
-
-    def post(self, student_id, course_id):
+    # @is_admin
+    def post(self, student_id):
         """
             Register a student for a course using student_id and course_id
         """
         
-        # get the student
+        current_user = get_jwt_identity()
         student = Student.query.filter_by(id=student_id).first()
         if not student:
-            return {'message': 'Student does not exist'}, HTTPStatus.NOT_FOUND
-        
-        # get the course
-        data = request.get_json()
-        course = Course.query.filter_by(id=course_id).first()  
-        if course is None:
-            return {'message': 'Course does not exist'} , HTTPStatus.NOT_FOUND
-            #check if student is already registered for this course
-        get_student_in_course = StudentCourse.query.filter_by(student_id=student.id, course_id=course.id).first()
-        if get_student_in_course:
             return {
-                'message':'Course has already been registered'
-                } , HTTPStatus.OK
-        # Register the student to the course
-        add_student_to_course = StudentCourse(student_id=student.id, course_id=course.id)
-        add_student_to_course.save()
-        return {
+                'message': 'This student does not exist'
+            }, HTTPStatus.BAD_REQUEST
+        
+        data = admin_namespace.payload
+        new_course = StudentCourse(
+            course_code = data['course_code'],
+            course_unit = data['course_unit'],
+            score = data['score'],
+            grade = data['grade'],
+            first_name = data['first_name'],
+            last_name = data['last_name'],
+            student_id = student_id,
+            course_id = data['course_id']
+        )
+        new_course.save()
+
+        return new_course, HTTPStatus.CREATED, {
             'message': 'Student successfully registered for course'
-        }, HTTPStatus.CREATED
+        }
+
         
     
 
@@ -539,6 +517,7 @@ class GetStudentCourses(Resource):
         }
     )
     @jwt_required()
+    @is_admin
     def get(self, student_id):
         """
             Get all scores for a student
@@ -564,6 +543,7 @@ class CalculateGPA(Resource):
         }
     )
     @jwt_required()
+    @is_admin
     def get(self, student_id):
         """
             Calculate a student GPA
@@ -590,7 +570,28 @@ class CalculateGPA(Resource):
         
 
 
+@admin_namespace.route('/student/<int:student_id>')
+class GetStudentCourses(Resource):
+    @admin_namespace.marshal_with(course_list_model)
+    @admin_namespace.doc(
+        description='Get all courses a student registered for', params={
+            'student_id': 'The student id'
+        }
+    )
+    @jwt_required()
+    @is_admin
+    def get(self, student_id):
+        """
+            Get all courses a student registered for
+        """
+        student = Course.get_by_id(student_id)
+        if student:
+            return student.registered_courses, HTTPStatus.OK
+        else:
+            return student.registered_courses, HTTPStatus.OK
         
+
+
